requirements.txt
streamlit
pandas
plotly
python-telegram-bot
openpyxl
